XPS interpreter - fix 'synthetic bold' text style
Bug 696914 "Large black objects in output with pdfwrite device"

The problem here is that the XPS interpreter, when creating a synthetic
bold font, merely set the text rendering mode to 2 (fill and stroke),
assuming that the graphics library would deal with this.

However, the graphics library does not care about text rendering modes
at all. The graphics state parameter is used only by the high level
devices and (apart from this instance) only be the PDF interpreter. When
the PDF interpreter encounters a text rendering mode other than 0 it
checks the device to see whether it wants text rendering modes preserved.
If it does, then the interpreter simply sets the mode and does nothing
else. If, however, the device does not want the mode preserved, then the
PDF interpreter breaks the text rendering mode into its component
operations.

This commit adds the same functionality to the XPS interpreter. It seems
that the XPS interpreter only uses this for synthetic bold, and so only
uses Tr 2. So the code now checks with the device to see if it wants
the mode preserved, if it does, we simply leave it (but see below).
Otherwise, we stroke the text, and then draw it normally (fill). This
means that the artificial bold text is now actually drawn bold.

There is one slight complication. When stroking a path the pdfwrite device
takes care to undo the scaling done by the CTM from the stroke width.
But when using a text render mode, it does not do so, it uses the
line width 'as is'. The XPS interpreter sets the line width including
the CTM (which is correct for stroking), so we need to 'undo' the CTM
applied to the line width before we draw the text, when we are preserving
the text rendering mode.