Add support for -sPostRenderProfile="icc profile" to tiffscaled devices
With this commit, it is possible to have the tiffscaled32, tiffscaled24
and tiffscaled8 devices apply an ICC profile after rendering, which
will transform the rendered page to the color space defined by
PostRenderProfile.   When used in conjunction with -dUsePDFX3Profile,
if the source file contains an Output Intent profile, the page will
be rendered first to the color space defined by the output intent profile and then
transformed to the color space defined by the PostRenderProfile setting.  In this
way, we can avoid issues caused by mismatching ICC profiles in overprint
situations.  This along with the commit http://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=0df325b4bdbd037c92e2528fc16900de84f8d9c1
are needed to create a patch.