txtwrite - fix a conversion from bytes to shorts
Bug 696935 "SEGV in txtwrite"

When I modified the ToUnicode processing to return more than a single
code point for a given glyph, I altered the code to return a required
buffer size (or count of bytes copied). Txtwrite wanted a count of code
pints (always 2 bytes for ToUnicode) and so I should have divided the
return value by the size of a short. For some reason I multiplied it
instead, I have no idea what I was thinking of.....

I couldn't reproduce a SEGV but since this was a buffer overrun problem
it would depend on the memory layout. With this fix the text is once
again sensible, and Robin reports the SEGV has gone away.