Bug 696837: Improve performance of memory searches.
The swap to use splay trees reduces the performance of this
file from 11 seconds to 25 seconds for me.

This is due to the additional complexity of traversing a
more complex memory structure. The hope is that in the average
case we'll do better (or at least no worse), and that we'll
make gains in more pathological cases.

It turns out that in alloc_is_since_save I was failing to
actually make use of the fact that the new structure is
sorted and can hence be searched much more efficiently.

Do that here. This gets performance back to 17 seconds - so
still a net loss, but less so.