Use ToUnicode for substituted CIDFonts
When we substitute a CIDFont in a PDF file with a TTF from disk, use the
ToUnicode CMap (if available) to extract Unicode values. In a lot of case, this
gets a more accurate character code to glyph mapping than the previous reliance
on the low level glyph ordering from the TTF.

This also meant being more robust for broken ToUnicode CMap. Also, an additional
option (-dIgnoreToUnicode) to skip parsing of the ToUnicode CMap - for cases
where a correctly formed ToUnicode is semantically wrong.