Bug 696937: readhexstring handle signed/unsigned char
When readhexstring runs out of data, we overload an integer ref with the
number of bytes from the input buffer used, and the value of a trailing
odd numbered byte shifted to the top 8 bits of the integer value. If no
trailing byte was read, the value is set to -1.

When restarting readhexstring with a full buffer, we relied on casting to
char to retrieve the signed value from the top 8 bits on the integer ref.

Unfortunately, unqualified chars are not signed on all platforms, and on
platforms where chars are unsigned, we ended up with an invalid value.

So, have the casting use qualified (signed/unsigned) chars for writing and
reading back the value.

To make this neater, and consistent, add an "schar" type to match the existing
"uchar" type.