PDF interpreter - treat invalid BaseEncoding as a missing BaseEncoding
Bug #696942 "Missing content reading PDF file"

The file contains Encoding dictionaries which contain BaseEncoding
entries which are invalid, one is '/utf-8' the other is '/'. We already
checked for '/' and ignored it, and we already checked named Encodings
against the permissible list of Encodings, so here we just extend the
checks for BaseEncoding to see if a named BaseEncoding is in the list
pf permitted names.

No differences expected.