High level devices - handle 12 BPC image data
Bug #696944 "Lockup with pdfwrite device"

This appears to be a bug of very long standing and points up our previous
lack of image test files in the cluster runs, since it hasn't emerged
before.

The high level image stream code did not properly handle 12 bit data,
In s_compr_chooser__unpack_and_recognize() we ended up with 8 bits of
data remaining, which is not less than 8, but because the input size is
12 bits it is also not > the bits per sample. This led to an infinite
loop trying to read data at the end of a line.

Here we change the '<' to '<=' to properly detect the end of input, so
we don't fall into this trap of assuming samples are 8 BPC.