PDF interpreter - handle PageLabels where the number tree starts with Kids
Bug #696947 "Regression: Error: /undefined in --run-- writing pdf file starting with 5784bfbfba7191cacce5309e88afac0851287460"

This is not really a regression, exactly. Prior to the commit noted
above we discarded PageLabels, the commit added the ability to preserve
these.

However, we had a limited number of available files to test from, and
this particular file is the first one I have encountered where the
number tree defining the page ranges begins with a number tree node
which has /Kids but no /Nums (Table 3.34, p166 of the 1.7 specification
says that the Root node may have /Kids only if there are no /Nums)

I did code to cope with this but there was a minor flaw in the code,
fixed here, which we lacked a test case for.

Note that, although the specification says on p595 that the value of
the PageLabels entry in the Catalog is a number tree, this file neatly
exposes a probable bug in Acrobat. If the root of the number tree has
no /Nums, but does have /Kids (which is valid for a number tree, see
above) then Acrobat X does not display the PageLabels.

This probably explains why we have never seen such a file before.

For now I have chosen to permit the number tree to have /Kids, if in
the future we need to revise this to behave like Acrobat we can do so.

This does have the interesting side effect that if the file is sent to
the pdfwrite device (and there is no other device which can use
PageLabels) the resulting PDF file will have working labels, where the
original does not.

No differences expected