Fix DeviceGrayToK. Bug 696875
By default we have -dDeviceGrayToK=true map all gray source
colors directly to the K channel when the Device is CMYK based.
This fix changes it so that Ghostscript matches Acrobat and
Distiller and does this mapping only if the source color
was truly defined as DeviceGray.

The file in Bug 696875 has the property that it has a gray image
whose color space is DefaultGray, which references an ICC color
space.  This color space was being treated as DeviceGray by
Ghostscript but it really should be color managed and not
treated as DeviceGray since it is ICC based.

There are several progressions in the test suite assuming the
reference is Acrobat.  The differences are caused by the fact
that gray sources that are not DeviceGray are now color
managed.