Fix memory leak due to PDF interpreter DefaultQstate losing scope
The graphics state is mostly tracked during gsave/grestore operations,
but the 'gstate' operator can result in a graphics state not being
freed until a garbage collection. Add a 'finalization' function for
gstates to deal with this.

The alloc_restore_all function needs to do gs_grestoreall_for_restore
so that the gstates don't end up with dangling pointers, so pass the
interpreter context instead of the dual memory pointer so it can call
that function.

Thanks to Chris for coming up with this approach to the 'finit' freeing,
and finding the fix for a SEGV due missing rc_init in gx_cie_to_xyz_alloc
that affected a couple of PS FTS files.