Skip 0 length edges when filling.
In theory, a zero length edge shouldn't make any difference
to the appearance of a path. In practice, they appear as
horizontal edges, and so can trip the code in gxfilltr.h that's
labelled as:

     /*
      * This is a hack to make sure that isolated horizontal
      * lines get stroked.
      */

That code is only called for !PSEUDO_RASTERIZATION.

There is code in gxfill.c that ignores zero length edges, but
this is currently conditional on pseudo_rasterization being
turned on.

As I understand it, PSEUDO_RASTERISATION is 1 when we are
plotting chars, and should never be so anymore. The idea is that
PSEUDO_RASTERISATION prevents dropouts in characters, hence it
feels wrong that any code that has this on should plot LESS than
code with it off.

I suspect that Igor got the test the wrong way around here.

I am just removing the test, and always ignoring 0 length edges.