Fix FirstPage/LastPage for non-PDF input
Robin spotted this one. When I added the page ranges code to augment
the existing First/Last page functionality I moved the code which
decided whether an operation was destined for a page which was being
output into a single routine. Previously this was simple code which
was replicated in each method, but the added complexity meant it was
better handled centrally.

However, when doing so I accidentally dropped the '-1' from the First
and Last Page tests (PageCount starts from 0) which meant these were
off by one throughout.

Added these back in here.