pdfwrite - don't coerce Dests from strings to names unless required
Bug #696974 "GS pdfwrite seems to break /GoToR's destination names /D"

We were always altering GoTo and GoToR Destinations from strings into
names, because string Dests are a PDF 1.2 feature. However its not
always possible to represent a string (eg Unicode) as a name.

In this commit we don't try to force the Dest into a name unless the
PDF level is < 1.2 and if we do try to coerce it, we check to see if it
can be represented. If not we throw an error and drop the Dest.