PosScript interpreter - don't error on nested forms
In commit fa20f5915978823a8c72a80e49fa90ce9c5c5879 I added code to check
if a Form PaintProc illegally left junk on the stack after execution
and cleaned it up if it did (issuing a warning).

However, the code didn't cease checking when it found the first Form
dictionary, which meant that if there were two Form dictionaries on the
stack (nested forms) then we would throw an error.

This commit simply terminates the loop when we find and check a Form
dictionary. If forms are nested then terminating the enclosing form will
lead it to check its own dictionary on the stack and so on.