pdfwrite - permit nested Forms in PDF output
Bug #696986 "Add support for nested PostScript forms as nested Form XObjects in pdfwrite"

Previously we checked for nested forms and 'unrolled' the child form(s)
into the parent PaintProc. However a simple extension (declaring the
child Form XObject(s) as Resources for the parent) permits us to have
nested forms.

Note that the way our PostScript output (from ps2write) works, it is
not possible to support nested forms, so in this case we continue to
unroll the child PaintProc into the parent. The code works, despite
the inefficiency.