Fix global/local VM issue with gstate that b9a265a trips over.
The commit to perform finalization cleanup of gstates had problems at
the alloc_restore_all when global memory was finalized because local
memory was already freed, but there was still "savedinitalgstate" in
systemdict that was in global VM but that had elements that were in
local VM.

Also .forceundef savedinitialgstate in systemdict to avoid leaving
a pointer to localVM from systemdict. This isn't strictly needed
since systemdict isn't ever traced by restore_finalize.