Scan converter configuration code.
Here, we introduce the notion of "which scan converter are
we using" to the core library. This is an integer value
in gs_lib_ctx_t which (currently) defaults to 1.

We add graphics library calls to get/set this value, and
new operators reflecting these into postscript.

In addition, we update both the PCL frontend and the PS init
code so that -dSCANCONVERTERTYPE can be used on the command
line to set the desired scan converter value.

-dSCANCONVERTERTYPE or -dSCANCONVERTERTYPE=1 will set it to
1. -dSCANCONVERTERTYPE=false or -dSCANCONVERTERTYPE=0 will
set it to 0. Otherwise -dSCANCONVERTERTYPE=<int> will set
it to the required value.

Currently we only have the one scan converter in gs, hence
setting this value makes no difference. A subsequent commit
(probably the very next one) will introduce an alternative
scan converter that will be disabled by default, but can be
enabled by setting this value. This commit merely prepares
the framework for that to fit into.

The intent is that:

  0 (or lower) will mean "The old scan converter"
  1 will mean "The default scan converter"
  2 (or higher) will mean new scan converters yet to be defined.

Thanks to Ken for the postscript code.