Update PCL command line parser arg handling.
Certain "special" params are interpreted directly by the pcl
build, and handled directly. (Namely, BATCH, NOPAUSE,
NOINTERPOLATE and NOCACHE).

The current code can only cope with these being being used
with no argument. i.e. -dNOINTERPOLATE will be accepted, but
-dNOINTERPOLATE=1 will not trigger the no interpolation code.

Here we change the code so that we now decide whether the arg is
a 'special' one or a normal param at the last moment, and
correctly raise errors if the values are set out of line.