Decrement rather than free reference counted object
Commit b9a265a02b7d1 fixed the unintialised reference count data for the
cie_joint_caches in a temporary gstate created during CIE to ICC conversion.

I hadn't realised it was explicitly freed later on, rather than either
had the ref count decremented, or left to the default gstate freeing to deal
with. This was causing a C lib error on Windows, and a valgrind error on
Linux.

Switching to using the reference counting machinery resolves the problem.