Make 32-bit CMYK image icc-proccessed, unless UseFastColor is specified.
This is a further part of the enhancement for bug 696905.

There is a big block of comments about a previous change of
pclxl_can_handle_color_space() (introduced in bug 692329)
being wrong and causes slightly wrong colors. That needs to be
looked at eventually, and this change updated.

No cluster differences