Bit baffling this one. When running a pattern we use the graphics state extant at the time when then enclosing object was created. So for pages that's the page, when inside a form its the state at the time the form was executed and so on.
The pattern instance is created with a graphics state which is stored
as part of the pattern instance, and used when replying the PaintProc.

However, if the pattern is executed inside transparency, it seems we
do need to apply the 'current' transparency parameters before we
instantiate the pattern. Otherwise if we run a form inside a pattern
the form won't be created with the correct graphics state parameters
as it will inherit them from the 'enclosing' context, not the one
active at the time the pattern is drawn.