Allow 32-bit CMYK image (converted to 24-bit RGB) to be jpeg compressed.
Previously, anything except 24-bit RGB is forbidden from going through
jpeg compression, because anything other than image does not make sense.
RLE and DeltaRow can apply to general byte-stream data, not jpeg.

This is another part of the enhancement for bug 696905, and depends
on the previous two of this group. The previous two changes convert
32-bit CMYK data to 24-bit RGB internally, either through
DeviceCMYK->RGB or icc. So this change is just a one-liner.

No cluster differences