Bug 697056: Fix LCMS static init of critical section.
LCMS2 relies on being able to statically init a mutex
(actually a Critical Section on windows). Windows has
no official API for statically initing a critical
section, so lcms2 uses a hack (albeit it a safe one)
to do this.

Unfortunately ApplicationVerifier (and presumably other
similar tools) don't understand this idiom, and get
confused by not seeing an explicit initialisation call.
This is causing problems for a customer.

This commit therefore introduces code to properly
initialise it on Windows. In order to avoid any possible
multithreaded startup issues, we use the Windows
InterlockedCompareExchangePointer API (introduced in
Windows XP).

If anyone wants to avoid this code (so they can work
on pre-Windows XP), they can build with:

 CMS_RELY_ON_WINDOWS_STATIC_MUTEX_INIT

defined, and get the old behaviour. We automatically
set this when building with any version of MSVC older
than VS2005 (i.e. those that don't have that API
available).