Improve bug 696985. Performance issue (and large memory usage).
The GC would run due to the 'system' VM which never had it's limit set (default
600,000), but after the GC ran the limits for all of the VM's were reset to
the current allocated plus the limit, so the local VM and global VM were not
in a state to be collected (pages ran in a save ... restore).

Setting the system VM limit to the same value allows the file to complete
in 84 seconds and only uses 74Mb (a cut down file that was 5k pages took 170
seconds and used 634Mb before the change). The entire large file only needed
725 GC's to complete and most were for the local VM (space==12), rather than
for the system VM (space==4).

Also change the formatting of one of the gs_debug['0'] messages to keep all
the relevant values on one line which makes searching/grepping easier.