Add script and arch*.h headers for ios build
This results in a "fat" libgs.a library which can be linked used either for the
simulator or a "real" ios build.

Really, this script just automates/simplifies the process of using our existing
cross-compile support.

The only slightly out of the ordinary part is combining the Intel and arm
libraries into the single fat library.