Bug 696599: cast disguised wrong type dereference.
The code was casting to a gx_device_printer to get to the BLS_force_memory
value, but since the target device for a clist device is no longer
necessarily a printer device, it wasn't guaranteed that the struct contained
that entry.

So move BLS_force_memory to the base device type (gx_device), so it's always
available.