pdfwrite - Don't limit the resolution when Compatibility > 1.4
Bug #697098 "pdfwrite charpath conversion problem"

This is not a fix, but a work-around, especially for future files.

We currently limit the resolution by checking the media size to see if
it lies in the range +/- 16,000. This is because The PDF Reference states
that co-ordinates are real numbers, and versions of the spec prior to
1.5 have a limit of +/- 32,767 for real numbers. In practice it was
noted that Acrobat Reader couldn't even cope with numbers that large so
we arbitrarily settled on 16,384 and then allowed a bit of slop.

With the PDF 1.5 specification (now 13 years old) this limit was raised
to +/- 3.403x10^38

In the interim we have also raised our default PDF creation to 1.5 and
we have adopted other means (altering the CTM) to keep co-ordinates
in the 32,767 range. So there is really no reason to maintain this old
 hack.