Add same approach for 'strip' as for pkgconfig
where we use the AC_PATH_TOOL() to try to find the correct instance of the
exe.

If we are cross compiling, we also use AC_PATH_PROG() which only finds the
'normal' (non-cross compile toolchain) exe. We then check the results: if
both approaches yield the same exe, then assume that no cross-compile compatible
version is available, and do not try to use the executable.