pdfwrite - don't reset the graphics state object for new XObject streams
Bug #697109 "Inkscape graphics in Latex PDF distorted on compression"

When starting a new substream the pdfwrite code always resets the current
graphics state to the 'initial' state. This seem, to me, utterly wrong,
if we have changed the graphics state, then it persists into any
'substreams'.

Indeed, in this case we set the dash array, then the fact that we have
transparency causes us to push a new group, this starts a new substream,
we reset the graphics parameters (including the dash). Then we set the
dash pattern to the default and stroke a path. If we didn't start a
group we would notice the dash had changed and emit the change. However
because the group causes a reset, we see the dash change to the same as
it currently is, so we don't emit it.

I'm not totally happy with this change, there are a *lot* of subtleties
in the code, not least the (bonkers) fact that starting a group *does*
reset 3 extended graphics state parameters (SMask and constant alpha).
It looks like we have some assumptions built in elsewhere as well. If
I don't reset all the parameters then some other files start to fail.

This fix doesn't reset the graphics state for XObjects, but does for
everything else. For XObjects it only resets the 3 group parameters.

I'd spend more time on this but I want to get some sort of fix into the
9.20 release. This definitely needs more investigation.

There are a number of diffs, most are definite progressions, a few that
I can't be sure of but they don't seem any worse (and in some cases the
original file is technically broken anyway). Oddly there are more
progressions when rendering to halftoned devices, I can only assume
that we were previously emitting a broken halftone.