PDF interpreter - fix for GSView 5
Commit 093bd18bd923644fcd25c507088c0ebc63b4c320 included an
'optimisation' to prevent rescanning for transparency if we had already
scanned. However, this assumed that the function pdfshowpage_setup
had already been executed in order to set a variable.

This is always true when running files via Ghostscript, but any
application (such as GSView 5) which executes the PDF operations
individually need not execute this function. If the function was not
executed then an error occurred. From other comments in pdf_main.ps
its possible that customers may be using these functions as well (see
the comments above the definition of /pdfshowpage)

This commit checks to see if the variable has been set, if it has then
it is used, otherwise we rescan for transparency. This prevents the
error, but uses the optimisation if its possible to do so.

No differences expected