Fix bug 697097. SMask subpixel offset must match image offset.
The dda used to select the source pixel for mapping into the image
must use the same stepping for gray (monochrome) and color images
since an SMask is monochrome and the image may be in color. This is
primarily evident with the bug file since Matte is used to indicate
that the source data is premuliplied. Colors can shift a LOT since
the removal of the premultiplication can expand mistaked.