Bug 697087: Apply PDF TTF rules when loading from disk
When we substitute a TTF from disk for a (non-embedded) TTF in a PDF, apply the
same rules (if possible) as when reading a TTF embedded in a PDF. So, selection
of cmap table, font encoding, whether or not to draw the notdef etc.

This applies *only* to fonts, *not* to CIDFonts.