OpenJPEG allocation fixes.
The old version of openjpeg allocates using malloc/free.

The new version of openjpeg allocates via opj_malloc (and co).

We can capture these and pass them down to our own allocators,
but we cannot get openjpeg to pass us a gs_memory_t * to use.

We therefore resort to using a static to hold 'the current
gs_memory_t *' before we call openjpeg. To keep this threadsafe
we add some private data to the gslibctx, which the openjpeg
implementation can use to hold a lock.

We add calls to both jpx decode implementations to init this
private data (where the luratech implementation is null).