Avoid adding an extra makefile
Instead of haveng a special makefile to hold CCAUX and related values, have
configure create a temporary makefile for the cross compiling case (before
the recursive call to configure), and then pull the appropriate values from
that.

This means that the non-cross-compile case is simpler and closer to the previous
function.