Remove (and re-implement) ConvertUTF.c
Bug #697122 " embedded ConvertUTF.c is buggy and licensed incompatibly with GPL/APGL"

Its not clear that this code is incompatible with GPL, nor do we think
any 'bugginess' in the code affects us, since we are using a comparatively
small part of the included code.

Nevertheless its possible to remove the code, and re-implement the small
part we actually need, and that is done here.

Also removed the DSCEncodingToUnicode option which was insanely difficult
to use, and incorrectly documented.

Yhis shows one difference, 692486_-_heap_overflow_in_pdf_to_ucs2.pdf
now correctly throws an error, because the PDF file contains document
information (Application) which has an invalid UTF16-BE sequence.