pdfwrite - Fix handling of Separation /All alternate space conversion
Bug #697118 "Separation colorspace is not converted correctly by pdfwrite"

The /All Separation space wasn't getting the components of the tint
transform colour space correctly set. For rendering we never need to
do that, and simply set the relevant component of the Separation
(or DeviceN) space to be the tint value.

This commit forces the CMS to run the tint transform when we need to
create a new, different, alternate for the Separation space.

No differences expected.