pdfwrite - fix ToUnicode CMap sequence consolidation
Bug #697176 "ToUnicode CMap generation broken in 9.20"

The reworking to handle the current version of ToUnicode CMap missed
a case where the size of a code pair had changed. This led to a loop
comparing incorrect values, instead of comparing two consecutive
values, it compared a value with one much earlier.

In general this led to a loss of efficient, causing us to emit two or
more entries when it was possible to emit a single range. However in the
rare case when the comparison happened to compare two values that were
consecutive, but shouldn't have been, this led to us writing an incorrect
ToUnicode CMap.

No differences expected.