Coverity ID 137283 - pdfwrite - remove some redundant code
I had left some code that catered for a UTF16 surrogate pair promoted to
UTF32, converting the result to UTF8. But I didn't implement the UTF32
promotion, leaving it as a potential future enhancement.

However, Coverity whinges about the test always being true, so here
just remove the code. Its trivial enough to add in if we ever need it.