Fix pdfmark argument validation
This doesn't fit neatly into any category, its a PostScript 'operator'
but it is only implemented for the pdfwrite device...

The pdfmark operator takes a mark, a series of arguments and a name
object. Zero argument sis valid, but the absence of a name object is
not. We weren't checking for the absence of a name object, or
validating its type.

No differences expected.