Fix Bug 696864: Use Default colorspaces when processing SMask
This is done by saving the OverrideICC user parameter before setting the
colorspace, etc. so as to ignore the ICC profile, which may have a gamma
that we don't expect, and restore it after the SMask has been rendered.
Note this file rendered correctly with -dOverrideICC on the command line,
but this patch automatically applies that during execmaskgroup.

Thanks to Michael Vrhel for identifying the problem and coming up with a
simplified test case.