pdfwrite - fix SMask handling, when painting with a Pattern
Bug #697212 "Error in SMask during ExtGState"

The title is misleading, the error occurs with a Pattern, but there must
be an SMask preceding it. This appears to have been a very long-standing
problem in the code, but it was hidden by ignoring an error return code.
Fixing Coverity reports meant that the return code was actioned, and that
is what revealed the problem.

When we start a new substream to capture a Pattern PaintProc we need to
reset the 'soft_mask_id' (the ID assigned to the SMask pdfwrite has
created) in the graphics state. If we don't do that then the code (in
effect) tries to handle the SMask twice, once for the correct stream
and once for the pattern PaintProc stream, and this results in the
stack of pdfwrite internal gstate copies becoming unsynchronised, as well
as creating a PDF file with unbalanced q/Q operators.

Ignoring the return code had masked this problem.

No differences expected.