Bug 692721: remove SMask specific stuff from form dict
When we get an SMask group, we have to store the graphics state and the contents
of the ExtGState, and reset all that before we execute the content stream
for the group XObject. So it all gets setup in the correct place, that
information needs to be added to the group's form dictionary.

It turns out we also need to *remove* these entries from the group's form
dictionary, in case it gets reused as a normal for XObject.

The gstate and extended gstate values remain in the SMask parameter dictionary,
so any reuse of the SMask will still work correctly.