Rework openjpeg subsampled decode.
The current code performs the yuv -> rgb step on the data
within openjpegs buffers. Unfortunately, this means that
subsampled data is "expanded" into rgb, converted, and then
redecimated. The R plane stays intact, and the G and the B
are degraded. The code is then desubsampled again when copying
out into the stream, but the damage has been done.

This gives awful results.

Instead here, we simplify the code to decode a row at a time
into a row buffer. We then convert that row buffer to rgb, and
copy that out. We never redecimate the output.

This looks much better.