Bug 697144: Fix radialshade1.ps
When filling a radial shading, we form 4 tensor patches, one for
each quadrant. The points for these are represented as poles of
the tensor patches as p->poles[v][u].

One quadrant might look like (excuse the ascii art)

  J I H G          p->poles[0][0] = A   p->poles[3][0] = J
        F          p->poles[0][1] = B   p->poles[2][0] = K
  K      ED        p->poles[0][2] = C   p->poles[1][0] = L
                   p->poles[0][3] = D
          C        p->poles[1][3] = E   (plus 4 more points
	           p->poles[2][3] = F    internal to the
          B        p->poles[3][3] = G    patch, not shown
	           p->poles[3][2] = H    here)
       L  A        p->poles[3][1] = I

Thus 'u' controls how far along the radius of the radial shading
we are, and 'v' controls how far 'around' the shading we are.

The shading code first subdivides the patch on v (using the
split_patch routine). This gives us a succession of slices running
from the centre of the radial fill to the outside (like slices of
pizza).

It then subdivides these slices (using split_slice) to give small
quadrangles to fill.

The code then checks to see whether these quadrangles need to be
further divided by checking to see whether the color change across
such quadrangles is monotonic or not. If it is, and the quadrangles
are 'not big' (for some definition of big) then it is considered
reasonable to fill them using a linear fill.

For simple radial fills, the color changes across 'u' not 'v'.
We thefore check for the color change in the given range to see
if it is monotonic, and if it not, we divide the patch.

As part of the work for bug 689189 (see commit 36375961a02091)
Igor had identified some uses of the radial fill code where the
tensor patches are constructed such that the color varies across
'v' rather than 'u', and had therefore amended the code so allow
for this.

Unfortunately the setup for this appears to have got it wrong
for the normal radial case. Fixing that here solves the problem.