Bug 697345 Blend Color Space Support for Separation devices
This is a rather large commit that brings support for transparency
blend color spaces to the separation devices.  Previously the
transparency compositor always used CMYK for the blend color space
if the output device was a separation device.

With this commit:

If the blend space is RGB or Gray based, then we now ensure that the
alternate tint transform is not used when we encounter a separation or DeviceN color space.
(assuming we have not run out of spot color space at the target device).   Note that
if the any of the spot colors in a DeviceN color space are CMYK process colorants and the
blend space is Gray or RGB, the alternate tint transform IS used.

2) The pdf14 compositor now handles a mixture of additive and subtractive components.
I.e. RGB + spots or Gray + spots.

3) If the blend mode is non white preserving or not separable, then the spot colors
use the normal blend mode while the process colorants use the specified blend mode.

4) In the process there was a bit of code clean up. But much remains to be cleaned.