Bug 697144: Fix radialshade2.ps
Many many moons ago, a fix was committed for bug 690831. The
commit message for that was lost in the great svn -> git
migration, but the bug itself makes the history clear.

Looking back at this again now, it appears that the fix was
incorrect. Ghostscript was in fact doing the right thing by
the letter of the standard, but Acrobat was applying a fudge
factor.

When filling a radial shade, the code tests to see if the
two circles are nested. If so, then when the shade is 'extended',
the entire area outside the final circle is filled. If not,
a different fill operation has to happen.

The file given in the bug has a shade where the two circles are
very close to being nested, but are not strictly so. Acrobat is
choosing to treat them as if they were.

We therefore revert the code to 'unbreak' it, and adjust the
fudge factor in our code to give the same results as Acrobat does.

This causes no changes in the cluster tests.