Bug 697144: Fix radial shadings broken by previous fix.
The previous fix addressed the fact that monotonicity of a
filled quad was being tested based on splitting on u rather
than v (or vice versa). It turns out that in some
circumstances (when circles are entirely nested), we reverse
u and v, and hence need to reverse the test.

The previous fix did not allow for this.