Bug 697220(1): Push systemdict onto dict stack before we cleanup.
As part of shutting down the Postscript interpreter, we run some Postscript
from the C code. We really don't want to stumble across random redefinitions
of operators (especially if the definitions trigger errors) at this point.

So, push systemdict onto the top of the dict stack before we start our cleanup.

Do it in C, to further ensure we don't trip over redefinitions in the Postscipt
world.

This *should* resolve the segfaults, but leaves a garbage collector error