Bug 697231: Introduce caching to use of clipping paths.
When we create a clipping device from a path, we copy over the
pointers to the rectangle list, and start searching them from
the start.

This means that if we perform an operation that causes lots of
calls to something with the same clipping path (like for instance
a stroke operation with a clipping path that calls down to a fill
path with a clipping path), then we will repeatedly search from
the top of the rectangle list.

This massively slows stuff down.

This commit introduces a 'cached' entry in the clipping path
that can be updated after a clipping device has been used to show
the last place in the list that was found. Subsequent creations
start searching from this point.

This gives us the desired locality of reference, and massively
speeds up the test file.